refactor(translator): extract legacy body path and error helpers
- Move full-body/chunked fallback into `legacy-body-translation.service.ts`
- Thread per-file LLM state through `TranslationFileContext`
- Classify segment-batch failures in `translator-error.helpers.ts`
- Skip legacy fallback on auth, quota, or non-splittable errors
