+++ src/app/services/translator/legacy/index.ts
@@ -1,0 +1,1 @@
+export * from "./legacy-body-translation.service";
